src/functions/reporting/requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
flask
